requirements.txt
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
docx2txt==0.8
futures==3.0.5
grpcio==1.64.0
grpcio-tools==1.64.0
idna==3.7
lxml==5.2.2
protobuf==5.27.0
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
python-docx==1.1.2
requests==2.32.3
setuptools==70.0.0
striprtf==0.0.26
typing_extensions==4.12.0
urllib3==2.2.1
